pyproject.toml
[tool.poetry]
name = "maccabipediabot"
version = "0.14.4"
description = "MaccabiPedia Python Telegram bot"
authors = ["MaccabiPedia <maccabipedia@gmail.com>"]
packages = [{ include = "maccabipediabot", from = "src" }]

[tool.poetry.dependencies]
python = "^3.7"
maccabistats = ">=2.33.0"
requests = ">=2.20.0"
python-dotenv = ">=0.12.0, <0.13"
python-telegram-bot = ">=12.4.2, <13"
decorator = ">=4.4.2, <5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
run_bot = 'maccabipediabot.bot_main:run'


[virtualenvs]
create = true
in-project = true
